recorrect define name

+++ 05-SceneManager/TimeCustom.h
@@ -3,8 +3,8 @@
 class TimeCustom
 {
 public:
-	int startTime;
-	int limitTime;
+	ULONGLONG  startTime;
+	ULONGLONG  limitTime;
 
 public:
